UPDATE notice entity seperate eduNotice with MMNotice #lee

+++ Back-End/ssachedule/src/main/java/com/ssljjong/ssachedule/repository/UserRepository.java
@@ -13,7 +13,7 @@ import java.util.List;
 import java.util.Optional;
 
 
-public interface UserRepository extends JpaRepository<UserDomain, Integer> {
+public interface UserRepository extends JpaRepository<UserDomain, Long> {
 
     /**
      * * find UserDtoList by Track
